Split session methods
